MID-6319: added count functionality and related tests

+++ repo/audit-impl/src/main/java/com/evolveum/midpoint/audit/impl/LoggerAuditServiceImpl.java
@@ -187,7 +187,7 @@ public class LoggerAuditServiceImpl implements AuditService {
     }
 
     @Override
-    public long countObjects(ObjectQuery query,
+    public int countObjects(ObjectQuery query,
             Collection<SelectorOptions<GetOperationOptions>> options, OperationResult parentResult) {
         throw new UnsupportedOperationException("countObjects not supported");
     }
